Update index.html
This portfolio is a non-exhaustive list of a variety of data science project that Benita Diop worked on. The aim of this portfolio is to showcase the magnitude of her technical analysis skills. She is experienced in statistical packages such as tableau, R, MATLAB, Mathematica; database languages such as SQL and Hadoop; general purpose languages such as Python and C++ and Visual Basic. various software and machine learning algorithms.